requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
Pillow==10.0.0
supabase==1.0.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
Flask-Login==0.6.2
requests==2.31.0
email-validator==2.0.0
python-slugify==8.0.1